refactor familyTable row parser
